requirements.txt
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
pytweening==1.2.0
rubicon-objc==0.5.0
Werkzeug==3.1.3
